+++ Assets/FlashTools/Docs/CHANGELOG.md
@@ -1,5 +1,5 @@
 ###### Version 1.3.15
-* Add bitmap trimming
+* Add optional bitmap trimming
 * Fix preview leaks in the Editor mode
 * Add warning notes about outdated assets
 * Add log message about successfully converting